requirements.txt
Flask==3.1.0
sqlalchemy==2.0.36
requests==2.32.3
apscheduler==3.10.4
jinja2==3.1.4
python-dotenv==1.0.1
pytest==8.3.4
